feat:ユーザー情報についてのRepository周りを実装

+++ src/main/java/com/example/procdpchallenger/application/port/outbound/user/UserRegistrationRepository.java
@@ -6,6 +6,6 @@ import java.time.LocalDate;
 
 public interface UserRegistrationRepository {
     void save(UserForRegistration user);
-    long countRegistrationsByDate(LocalDate date);
+    int countRegistrationsByDate(LocalDate date);
     int countRegistrations();
 }
